% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/killJob.R
\name{cancelJob}
\alias{cancelJob}
\title{Cancel a Postgres job.}
\usage{
cancelJob(src, pid)
}
\arguments{
\item{src}{It can either be a cdm_reference, a postgres_source or a
PqConnection object.}

\item{pid}{Numeric. The pid to cancel, multiple can be supplied.}
}
\value{
Invisible TRUE if the process is successful.
}
\description{
Cancel a Postgres job.
}
