% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmFromPostgres.R
\name{cdmFromPostgres}
\alias{cdmFromPostgres}
\title{Create a <cdm_reference> object from a postgres connection}
\usage{
cdmFromPostgres(
  con,
  cdmName = NULL,
  cdmSchema = "public",
  cdmPrefix = "",
  writeSchema = "results",
  writePrefix = "",
  achillesSchema = NULL,
  achillesPrefix = "",
  cdmVersion = NULL,
  cohortTables = character(),
  validation = TRUE
)
}
\arguments{
\item{con}{A PqConnection created with DBI and RPostgres.}

\item{cdmName}{String, name of the cdm object.}

\item{cdmSchema}{String, name of the schema for the OMOP CDM Standard tables.}

\item{cdmPrefix}{String, prefix leading the OMOP CDM Standard tables.}

\item{writeSchema}{String, name of the schema for the cohort and permanent
tables.}

\item{writePrefix}{String, prefix leading the cohort and permanent tables.}

\item{achillesSchema}{String, name of the schema for the achilles tables.}

\item{achillesPrefix}{String, prefix leading the achilles tables.}

\item{cdmVersion}{String, OMOP CDM version of the cdm object.}

\item{cohortTables}{Character vector, names of the cohort tables.}

\item{validation}{Logical, wether to validate the cdm object or not.}
}
\value{
A <cdm_reference> object.
}
\description{
Create a <cdm_reference> object from a postgres connection
}
