% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyCdmToPostgres.R
\name{copyCdmToPostgres}
\alias{copyCdmToPostgres}
\title{Copy a cdm (independently of its source) to a}
\usage{
copyCdmToPostgres(
  cdm,
  con,
  cdmSchema = "public",
  cdmPrefix = "",
  writeSchema = "results",
  writePrefix = "",
  achillesSchema = NULL,
  achillesPrefix = ""
)
}
\arguments{
\item{cdm}{A <cdm_reference> object.}

\item{con}{A PqConnection created with DBI and RPostgres.}

\item{cdmSchema}{String, name of the schema for the OMOP CDM Standard tables.}

\item{cdmPrefix}{String, prefix leading the OMOP CDM Standard tables.}

\item{writeSchema}{String, name of the schema for the cohort and permanent
tables.}

\item{writePrefix}{String, prefix leading the cohort and permanent tables.}

\item{achillesSchema}{String, name of the schema for the achilles tables.}

\item{achillesPrefix}{String, prefix leading the achilles tables.}
}
\value{
A cdm reference inserted in the new pq_cdm source.
}
\description{
Copy a cdm (independently of its source) to a
}
\examples{
#library(omock)
#library(OmopPostgresConnector)
#
#pq <- localPostgres()
#cdm <- mockCdmFromDataset(datasetName = "GiBleed")
#
#pq_cdm <- copyCdmToPostgres(cdm = cdm, con = pq)
#pq_cdm

}
