% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{createIndexes}
\alias{createIndexes}
\title{Title}
\usage{
createIndexes(x, schema, table, index)
}
\arguments{
\item{x}{A PqConnection, cdm_reference, cdm_table or pq_cdm object.}

\item{schema}{schema}

\item{table}{table}

\item{index}{index to create}
}
\value{
The same x object.
}
\description{
Title
}
