% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{createTableIndexes}
\alias{createTableIndexes}
\title{Title}
\usage{
createTableIndexes(table)
}
\arguments{
\item{table}{A <cdm_table> object.}
}
\value{
The same <cdm_table> object.
}
\description{
Title
}
