% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{existingCdmIndexes}
\alias{existingCdmIndexes}
\title{Identify the existing indexes in the tables of a <cdm_reference> object.}
\usage{
existingCdmIndexes(cdm)
}
\arguments{
\item{cdm}{A <cdm_reference> object.}
}
\value{
A tibble object with the existing indexes.
}
\description{
Identify the existing indexes in the tables of a <cdm_reference> object.
}
