% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{existingTableIndexes}
\alias{existingTableIndexes}
\title{Title}
\usage{
existingTableIndexes(table)
}
\arguments{
\item{table}{A <cdm_table> object.}
}
\value{
A tibble object with the existing indexes.
}
\description{
Title
}
