% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{getJobs}
\alias{getJobs}
\title{Get ongoing Postgres jobs.}
\usage{
getJobs(src, user)
}
\arguments{
\item{src}{It can either be a cdm_reference, a postgres_source or a
PqConnection object.}

\item{user}{Users to filter by. If NULL no filter is applied.}
}
\value{
Tibble with the identified jobs.
}
\description{
Get ongoing Postgres jobs.
}
