% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmopPostgresConnector-package.R
\docType{package}
\name{OmopPostgresConnector-package}
\alias{OmopPostgresConnector}
\alias{OmopPostgresConnector-package}
\title{OmopPostgresConnector: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-pharmacoepi.github.io/OmopPostgresConnector/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
