% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresSource.R
\name{postgresSource}
\alias{postgresSource}
\title{Create a postgres source object}
\usage{
postgresSource(
  con,
  cdmSchema = "public",
  cdmPrefix = "",
  writeSchema = "results",
  writePrefix = "",
  achillesSchema = NULL,
  achillesPrefix = ""
)
}
\arguments{
\item{con}{A PqConnection created with DBI and RPostgres.}

\item{cdmSchema}{String, name of the schema for the OMOP CDM Standard tables.}

\item{cdmPrefix}{String, prefix leading the OMOP CDM Standard tables.}

\item{writeSchema}{String, name of the schema for the cohort and permanent
tables.}

\item{writePrefix}{String, prefix leading the cohort and permanent tables.}

\item{achillesSchema}{String, name of the schema for the achilles tables.}

\item{achillesPrefix}{String, prefix leading the achilles tables.}
}
\value{
A <pq_cdm> source object
}
\description{
Create a postgres source object
}
