% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{readPostgresLog}
\alias{readPostgresLog}
\title{Create a tibble from the log files in the logPath folder}
\usage{
readPostgresLog()
}
\value{
A tibble with the one row per log file in the logPath folder.
}
\description{
Create a tibble from the log files in the logPath folder
}
