% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{startPostgresLog}
\alias{startPostgresLog}
\title{Start a postgres log}
\usage{
startPostgresLog(path, sql = TRUE, explain = TRUE, analyse = FALSE)
}
\arguments{
\item{path}{Path to a folder where to store the log files.}

\item{sql}{Whether to log the sql information.}

\item{explain}{Whether to log the explain infotmation.}

\item{analyse}{Whether to log the analyse information.}
}
\value{
Invisible path.
}
\description{
Start a postgres log
}
