% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{tableDoc}
\alias{tableDoc}
\title{Helper for consistent documentation of \code{table}.}
\arguments{
\item{table}{A <cdm_table> object.}
}
\description{
Helper for consistent documentation of \code{table}.
}
\keyword{internal}
